requirements.txt
annotated-types==0.7.0
anyio==4.5.2
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.8
h11==0.14.0
idna==3.10
pillow==10.4.0
pydantic==2.10.6
pydantic_core==2.27.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.44.0
typing_extensions==4.12.2
uvicorn==0.33.0
